requirements.txt
# Location Intelligence Service Dependencies
# File: location-intelligence/requirements.txt

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
pgvector==0.2.4

# Location & Geocoding
geopy==2.4.1
shapely==2.0.2
pyproj==3.6.1

# HTTP Clients
httpx==0.26.0
requests==2.31.0

# Caching
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1

# Logging & Monitoring
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
